python/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
azure-data-tables==12.4.4
datetime==5.4